Add acoustic material model: mass-law, coincidence, and resonance
Extend the key-value material model with three acoustic mechanisms, wired
consistently through ray_mesh_interact and calc_diffraction_gain, and add
EM-safe tests. EM behavior is unchanged: new columns are absent for EM
materials, so every new term defaults to an exact no-op.

Helpers (quadriga_lib_helper_functions.hpp):
- medium_loss_dB: add guarded mass-law term max(0, m*log10((f/fRef)*dist));
  guard skips the m==0 case to avoid 0*log10(0)=NaN and zero-path artifacts.
- interface_loss_dB / interface_gain: new lumped per-entry transmission gain
  (att/attB power law plus optional Lorentzian coincidence dip), mirroring
  medium_gain. Factored so ray_mesh_interact and calc_diffraction_gain share it.
- eta_resonance: complex Lorentz permittivity pole (+i denominator, loss =
  negative imaginary part, consistent with the conductivity term). Applied to
  the Fresnel permittivity only, never to the real-sqrt bulk-loss path.
- medium_gain_impl: resolve and pass the "m" column.
- mtl_col: take a pointer (matches the _impl call sites).

ray_mesh_interact:
- Resolve m / resF / resQ / resS columns; per-material kR/kS reads.
- Split base permittivity (eta_med, in-medium loss) from resonant permittivity
  (eta1/eta2, Fresnel) so the Lorentz pole cannot push Re(eta)<0 into the
  bulk-loss sqrt. Pass mass to the two medium_loss_dB calls.
- Fix segfault: guard interface_gain_impl calls against null mtl_prop.

calc_diffraction_gain:
- Resolve res* columns in transition_gain_linear; fold the resonance into the
  reflection permittivity (consistent with ray_mesh_interact).
- Gate the dense->light reflection skip with `|| scalar_mode` so the scalar
  reflection path matches ray_mesh_interact's `&& !is_scalar` override gating.

Docs (format_materials.md):
- Rewrite for the name-keyed csv_prop model (drop the fixed 16-column matrix).
- Correct the resonance sign, clarify att (applied once on entry), in-medium
  loss composition (sigma + alpha + mass), scalar/no-TIR reality, and the
  per-interface 1-|R|^2 accounting; add a layered-materials (M2M) section.

Tests (test_acoustic_interactions.cpp):
- New acoustic-domain suite with Hz->GHz frequency mapping: mass-law slab
  traverse, coincidence dip, resonance absorption peak, EM<->acoustic
  convergence at both edges, and a ray_mesh_interact vs calc_diffraction_gain
  pass-through calibration on a large thin slab.
- Disable the obsolete EM-vs-scalar coord invariant in the diffraction tests
  (no longer holds once scalar applies reflection at oblique incidence).

+++ tests/mex_tests/test_point_inside_mesh.m
@@ -32,8 +32,8 @@ function test_point_inside_mesh
     % ───────────────────────────────────────────────────────────────
     res = quadriga_lib.point_inside_mesh(points, mesh, obj_ind);
     assertEqual(2, numel(res), 'Number of results must be 2');
-    assertEqual(uint32(2), res(1), 'First point should map to object index 2');
-    assertEqual(uint32(0), res(2), 'Second point is outside → 0');
+    assertEqual(uint64(2), res(1), 'First point should map to object index 2');
+    assertEqual(uint64(0), res(2), 'Second point is outside → 0');
 
     % ───────────────────────────────────────────────────────────────
     % Case 2: Provide only (points, mesh), no obj_ind, no distance
@@ -43,8 +43,8 @@ function test_point_inside_mesh
     % ───────────────────────────────────────────────────────────────
     res = quadriga_lib.point_inside_mesh(points, mesh);
     assertEqual(2, numel(res), 'Number of results must be 2');
-    assertEqual(uint32(1), res(1), 'Without obj_ind, returns face index+1 = 1');
-    assertEqual(uint32(0), res(2), 'Outside → 0');
+    assertEqual(uint64(1), res(1), 'Without obj_ind, returns face index+1 = 1');
+    assertEqual(uint64(0), res(2), 'Outside → 0');
 
     % ───────────────────────────────────────────────────────────────
     % Case 3: Provide (points, mesh, [], distance = 0.12)
@@ -55,8 +55,8 @@ function test_point_inside_mesh
     % ───────────────────────────────────────────────────────────────
     res = quadriga_lib.point_inside_mesh(points, mesh, [], 0.12);
     assertEqual(2, numel(res), 'Number of results must be 2');
-    assertEqual(uint32(1), res(1), 'First point within 0.12 → face index+1 = 1');
-    assertEqual(uint32(1), res(2), 'Second point within 0.12 of face 0 → 1');
+    assertEqual(uint64(1), res(1), 'First point within 0.12 → face index+1 = 1');
+    assertEqual(uint64(1), res(2), 'Second point within 0.12 of face 0 → 1');
 
     % ───────────────────────────────────────────────────────────────
     % Case 4: Provide (points, mesh, obj_ind, distance = 0.09)
@@ -66,8 +66,8 @@ function test_point_inside_mesh
     % ───────────────────────────────────────────────────────────────
     res = quadriga_lib.point_inside_mesh(points, mesh, obj_ind, 0.09);
     assertEqual(2, numel(res), 'Number of results must be 2');
-    assertEqual(uint32(2), res(1), 'First point exactly inside → obj_ind=2');
-    assertEqual(uint32(0), res(2), 'Second still outside → 0');
+    assertEqual(uint64(2), res(1), 'First point exactly inside → obj_ind=2');
+    assertEqual(uint64(0), res(2), 'Second still outside → 0');
 
     % ───────────────────────────────────────────────────────────────
     % Case 5: Provide (points, mesh, obj_ind, distance = 2.0)
@@ -78,6 +78,6 @@ function test_point_inside_mesh
     % ───────────────────────────────────────────────────────────────
     res = quadriga_lib.point_inside_mesh(points, mesh, obj_ind, 2.0);
     assertEqual(2, numel(res), 'Number of results must be 2');
-    assertEqual(uint32(2), res(1), 'First point → obj_ind = 2');
-    assertEqual(uint32(2), res(2), 'Second point within 2.0 → obj_ind = 2');
+    assertEqual(uint64(2), res(1), 'First point → obj_ind = 2');
+    assertEqual(uint64(2), res(2), 'Second point within 2.0 → obj_ind = 2');
 end
